+++ src/Azure.DevOps.PSModule/Tests/Pipeline/Check/Configuration/Get-AdoCheckConfiguration.Tests.ps1
@@ -24,7 +24,7 @@ Describe 'Get-AdoCheckConfiguration' {
                     }
                     settings = @{
                         approvers            = @(@{ id = 'user1' })
-                        definitionRef        = @{ id = 'def-id-1' }
+                        definitionRef        = @{ id = '8c6f20a7-a545-4486-9777-f762fafe0d4d' }
                         minRequiredApprovers = 1
                     }
                     resource = @{
@@ -42,7 +42,7 @@ Describe 'Get-AdoCheckConfiguration' {
                     }
                     settings = @{
                         allowedBranches  = @('refs/heads/main')
-                        definitionRef    = @{ id = 'def-id-2' }
+                        definitionRef    = @{ id = '86b05a0c-73e6-4f7d-b3cf-e38f3b39a75b' }
                     }
                     resource = @{
                         type = 'environment'
@@ -57,6 +57,9 @@ Describe 'Get-AdoCheckConfiguration' {
             id = 1
             name = 'TestEnvironment'
         }
+
+        $mockDefinitionRefApproval = @{ id = '8c6f20a7-a545-4486-9777-f762fafe0d4d'; name = 'approval' }
+        $mockDefinitionRefBranchControl = @{ id = '86b05a0c-73e6-4f7d-b3cf-e38f3b39a75b'; name = 'branchControl' }
     }
 
     Context 'Core Functionality Tests' {
@@ -209,6 +212,120 @@ Describe 'Get-AdoCheckConfiguration' {
 
             # Act & Assert
             { Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceName 'TestEnvironment' } | Should -Throw '*API Error: Unauthorized*'
+        }
+    }
+
+    Context 'ParameterSet Tests - ResourceId' {
+        BeforeEach {
+            Mock -ModuleName Azure.DevOps.PSModule Invoke-AdoRestMethod { return $mockCheckConfigurations }
+            Mock -ModuleName Azure.DevOps.PSModule Confirm-Default { }
+            Mock -ModuleName Azure.DevOps.PSModule Resolve-AdoDefinitionRef {
+                param($Name)
+                switch ($Name) {
+                    'approval' { return $mockDefinitionRefApproval }
+                    'branchControl' { return $mockDefinitionRefBranchControl }
+                    default { return @{ id = 'unknown-id'; name = $Name } }
+                }
+            }
+        }
+
+        It 'Should retrieve all check configurations using ResourceId without DefinitionType' {
+            # Act
+            $result = Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 2
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1
+        }
+
+        It 'Should retrieve filtered check configurations using ResourceId with DefinitionType' {
+            # Act
+            $result = Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345' -DefinitionType 'approval'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 1
+            $result.id | Should -Be 1
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1
+        }
+
+        It 'Should NOT call Get-AdoEnvironment when ResourceId is provided' {
+            # Arrange
+            Mock -ModuleName Azure.DevOps.PSModule Get-AdoEnvironment { return $mockEnvironment }
+
+            # Act
+            Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345'
+
+            # Assert
+            Should -Invoke Get-AdoEnvironment -ModuleName Azure.DevOps.PSModule -Times 0
+        }
+
+        It 'Should accept ResourceId via pipeline without DefinitionType' {
+            # Act
+            $result = [PSCustomObject]@{ ResourceId = '12345' } | Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 2
+        }
+
+        It 'Should accept ResourceId via pipeline with DefinitionType' {
+            # Act
+            $result = [PSCustomObject]@{ ResourceId = '12345'; DefinitionType = 'approval' } | Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 1
+        }
+
+        It 'Should work with ResourceId and multiple DefinitionTypes' {
+            # Act
+            $result = Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345' -DefinitionType 'approval','branchControl'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 2
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1
+        }
+
+        It 'Should work with ResourceId and Expands parameter without DefinitionType' {
+            # Act
+            $result = Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345' -Expands 'settings'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 2
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1 -ParameterFilter {
+                $QueryParameters -like '*$expand=settings*'
+            }
+        }
+
+        It 'Should work with ResourceId, DefinitionType and Expands parameter' {
+            # Act
+            $result = Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345' -DefinitionType 'approval' -Expands 'settings'
+
+            # Assert
+            $result | Should -Not -BeNullOrEmpty
+            $result | Should -HaveCount 1
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1 -ParameterFilter {
+                $QueryParameters -like '*$expand=settings*'
+            }
+        }
+
+        It 'Should construct correct URI with ResourceId in query parameters' {
+            # Act
+            Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceId '12345'
+
+            # Assert
+            Should -Invoke Invoke-AdoRestMethod -ModuleName Azure.DevOps.PSModule -Times 1 -ParameterFilter {
+                $QueryParameters -like '*resourceId=12345*'
+            }
+        }
+
+        It 'Should throw error when both ResourceName and ResourceId are provided' {
+            # Act & Assert
+            { Get-AdoCheckConfiguration -CollectionUri 'https://dev.azure.com/my-org' -ProjectName 'TestProject' -ResourceType 'environment' -ResourceName 'TestEnvironment' -ResourceId '12345' } | Should -Throw
         }
     }
 }